backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
proxmoxer==2.1.0
requests==2.32.3
httpx==0.28.1
python-jose[cryptography]==3.3.0
google-auth==2.37.0
slowapi==0.1.9
websockets==13.1
